Add experimental interactive debugging mode
In the event that a RUN command fails, when earth is run with the
interactive flag (-I); rather than clean up the failed build process, it
will present the user with an interactive shell.

This allows a user to manually inspect the system at the point of
failure, and additionally lets arbitrary commands be run interactively
without the need to modify the Earthfile.