Remove earthly-buildkitd container if exited
If the earthly-buildkitd container is in a stopped or exited state,
remove it before attempting to start a new instance of it.

This fix is to allow for the removal of the --rm flag from the
docker run command, so that logs persist in the cause of a failure.

This fixes https://github.com/earthly/earthly/issues/152