Remove extraneous checkout step in Actions example (#206)
* Remove extraneous checkout step in Actions example

GitHub Actions checkout@v2 no longer leaves the branch in a detached head state,
so there is no longer a need to run `git checkout` as a step.

* Remove checkout step from CI

Checkout@v2 no longer leaves the checkout in a detached HEAD state. This step
is no longer necessary.

The previous commit removed this block from the example.