Attempt ssh connection before falling back to https
Our previous heuristic for resolving auto auth was to check if ssh-agent
was running and contained a key; however, we've discovered at least one
user who had a valid ssh key loaded however that key was not configured
to use github which caused earthly's auto git protocol detection
heuristic to fail.

this changes the heuristic to first attempt an ssh connection, and if
that fails then default back to https.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>