prefix buildkit image with dev- when building +earthly
When version is not given, it defaults to the current git branch.
This causes a problem if someone checks out a previous branch of
earthly, modifies some code (while debugging a problem), and runs
earthly +earthly. This will overwrite the released buildkit image.

This PR adds a "dev-" prefix the buildkit when earthly is built
directly.

When performing a release via ./release+release, the VERSION tag is set
to whatever value is passed by RELEASE_TAG, (thus overridding the
default value (which includes the dev- prefix)).

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>