fix homebrew release process and instructions
fixes issue where PR was from `griswoldthecat/homebrew-earthly` was
opened against `griswoldthecat/homebrew-earthly` rather than `earthly/homebrew-earthly`

Update instructions to document how to apply the `pr-pull` label.
This label must be applied manually once the test checks complete and is
equivalent to approving the PR. For this reason this label can not be
automatically applied.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>