Prevent writes to directories above Earthfile
fixes a bug where a user could issue a

    SAVE ARTIFACT <dir> AS LOCAL .

which would overwrite the current working directory (including the
Earthfile). In a similar version of this bug, a user could suply ".."
which would delete the parent directory -- this could be repeated
all the way to the root directory (/).

Now when a user performs a

    SAVE ARTIFACT <path> AS LOCAL .

it will place `<path>` under `./<path>`.

This makes it more consistent with the current
non-local SAVE ARTIFACT commands which are all equivalent:

    SAVE ARTIFACT some-file .
    SAVE ARTIFACT some-file ./
    SAVE ARTIFACT some-file ./some-file

Futhermore, if a user tries to save to "../<somewhere else>", an error will
occur.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>