fix interactive tests
Now that we prevent running the --interactive flag when terminal is not
TTY; this caused the test from within interactive tests to fail to run.

This moves the docker-compose test to the python-based pexpect test
which now validates a PSQL query can be run within the postgres
container that is running inside docker-in-docker.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>