Always display frontend detection errors (#1382)
I had an error where `docker` was not on my PATH, and running earthly
displayed `auto frontend could not be initialized, but trying anyway`
without letting me know the issue. It wasn't obvious that a `-V` would
display the error message. Upon adding -V I got:
`auto frontend initialization error: failed to autodetect a supported frontend`.

This PR will now display the reason autodetect failed:

    auto frontend initialization failed due to failed to autodetect a supported frontend: 2 errors occurred:
        * docker-shell frontend failed to initalize: command failed: docker info --format={{.SecurityOptions}}: exec: "docker": executable file not found in $PATH: : exec: "docker": executable file not found in $PATH
        * podman-shell frontend failed to initalize: command failed: podman info --format={{.Host.Security.Rootless}}: exec: "podman": executable file not found in $PATH: : exec: "podman": executable file not found in $PATH
    ; but will try anyway

and in the event either of the frontends are detected, no error will be
displayed.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>