Allow setting env variables before starting buildkit
Introduces a new `buildkit_additional_entrypoint` global argument that
can be configured in `~/.earthly/config.yml`, which is evaluated before
running buildkit. This allows a user to pass arbitratry commands to the
entrypoint.sh script, e.g.

    buildkit_additional_entrypoint: 'export BUILDKIT_SCHEDULER_DEBUG=1'

or it can be extended to help with debugging, for example displaying all
environment variables containing the string debug:

    buildkit_additional_entrypoint: 'echo ---start---; env | grep -i debug; echo ---end---'

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>