allow overriding .env with --env-flag
Allow users to change which .env file is loaded by earthly; defaults
to .env in current directory.

To make this work, we need to get the value of --env-flag before
urfav/cli is initialized, as we must load in environment variables
into the global scope before parsing them with urfav/cli.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>