bugfix: COMMAND should run within cwd of caller
This fixes a bug where COMMANDs that defined a LOCALLY
would be executed within the directory of where the COMMAND
was defined rather than within the directory of the caller.

This new behaviour is consistent with the idea of a COMMAND
being like a macro where all statements from the COMMAND
are essentially inlined in the context of the caller.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>