Allow specifying cache size as percentage (implements #1723)
New config "cache_size_pct" (0-100) calculates a max cache size at buildkitd startup.  The percentage is relative to the available space on the filesystem containing the earthly data directory (which may be different from the space available on the host).  A value of 0 disables the setting.  If "cache_size_mb" is also present, the lower value wins.