Correct the default value of cache MB value
The assignment of the value is also moved to where other buildkitd settings are set.  This does slightly change the semantics around the deprecated flag, but hopefully for the better.  Other usages of this "flag trumps config" behavior have been removed over time as well, further supporting that the current behavior isn't intentional, just overlooked/forgotten.