New environment variables to control mirrors used by `dockerd`
This introduces `EARTHLY_DOCKERD_REGISTRY_MIRROR` and `EARTHLY_DOCKERD_INSECURE_REGISTRY` which allows control over the registries used by the `dockerd` instance used in `WITH DOCKER` commands.

Though buildkit is responsible for the majority of image pulling, docker (in docker) is responsible for pulling in some scenarios.

This allows for comma-seperated lists and will prepend an appropriate protocol if one is not present (to allow a buildkit-destined setting, which lacks protocol, to be used verbatim here).