Ability to set cache size limit as percentage of space (#1723) (#1734)
* Allow specifying cache size as percentage (implements #1723)

New config "cache_size_pct" (0-100) calculates a max cache size at buildkitd startup.  The percentage is relative to the available space on the filesystem containing the earthly data directory (which may be different from the space available on the host).  A value of 0 disables the setting.  If "cache_size_mb" is also present, the lower value wins.

* Documentation of new cache percentage setting

* Correct the default value of cache MB value

The assignment of the value is also moved to where other buildkitd settings are set.  This does slightly change the semantics around the deprecated flag, but hopefully for the better.  Other usages of this "flag trumps config" behavior have been removed over time as well, further supporting that the current behavior isn't intentional, just overlooked/forgotten.

* Add rudimentary tests for "cache_size_pct"

* Misc. link fixes and spelling/typo fixes

* Make output more readable/user-friendly

Co-authored-by: Alex Couture-Beil <alex@mofo.ca>

* Revert "Documentation of new cache percentage setting"

This reverts commit 6732892780fddf6f04e3c580d45f725dbe8d601d.

(moved to the `next` branch)

* Remove CLI flag for new cache percentage setting

The existing setting for cache MB has been deprecated, thus makes no sense to introduce one for this new setting

* Remove `--buildkit-cache-size-mb` CLI flag entirely

Previous commits resulted in new behavior (desirable or not), and was decided that it was time to be removed (having been deprecated for over a year): https://github.com/earthly/earthly/pull/1734#discussion_r824104125

* Refactor calculation of effective cache value

This simplifies the logic and messaging, though slightly more LOC.

One downside is that it no longer explicitly tells you the comparison being made, but anybody looking at this output would likely know they can easily run with `--debug` to see those values.

Co-authored-by: Alex Couture-Beil <alex@mofo.ca>