Introduce new "DOCKERHUB_MIRROR_INSECURE" environment variable for testing (#1733)
When set to "true" ("false" by default), a registry stanza is added that specifies that mirror "DOCKERHUB_MIRROR" is insecure.

The intended use-case is to support running a local caching/pass-through registry for use specifically for running tests and avoiding rate-limits (in the case of DockerHub) and/or reducing costly high-latency traffic (such as a corporate registry over VPN).

This setting is applied to the 'outer' buildkitd instance as well as the 'inner' instance.  The 'inner' instance now respects "DOCKERHUB_MIRROR" as well.