dind-alpine requires jq
jq is now required when using `WITH DOCKER` (dependency was introduced
in 88f94c5ed342a0162c202f5ce666d8350bf3fcc4), it is installed
automatically via `docker-auto-install.sh`.

dind-ubuntu already installs dependencies with this script, we should
use it for the alpine version as well to be consistent. Note that the
underlying docker:dind already provides dockerd, so only docker-compose
and jq will be installed.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>