Unit tests related to git clone functionality
These were mainly to help myself, but they're valid and work, so pushing these as well.  I didn't put a lot of time cleaning them up in case you didn't want them at all, or you wanted to go with a different solution.  If they're of value, great, but the separate commit is to aid in deletion.