WAIT/END support for save artifact as local
This extends the WAIT/END code to support waiting for an artifact to be
saved locally before continuing.

This introduces a new constraint that prevents a user from copying a
file into a context after it has already been output by earthly.
Previously such a workflow was not possible as outputs always happened
at the end of an earthly build (after all copies were performed).

This changes the save artifact as local workflow to use a random
alphanumeric string rather than an index-<number>, due to the fact that
multiple wait/end blocks may now be processed in parallel, which would
lead to conflicting index-<number> entries.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>