account password reset
This adds two new calls:

  earthly account reset --email <email>

which sends a request to the server for a password reset token to be
send to the supplied email.

and:

  earthly account reset --email <email> --token <token>

which then prompts for a new password, which is then sent to the server,
to perform a password reset.

This additionally fixes the password prompt to prevent displaying a
newline before the cursor that accepts a hidden password on stdin (this
mimics the behaviour of the sudo command).

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>