refactor builder.go to use wait_end.go
This introduces a new `--global-wait-end` flag which enables an implicit
global WAIT/END block, which the directly referenced target (and all
children) belong to.

This allows a large portion of the builder.go's build-function (bf), to
be skipped in favour of using code in wait_end.go' final `wait()` call.

This flag exists to give us test coverage for when we enable wait-end
blocks by default AND remove support for VERSION 0.6.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>