interactive debugger over unix socket
This changes the interactive debugger to use a unix socket rather than a tcp connection.
The unix socket is proxied over the buildkit session and is encrypted.

This simplifies the debugger as no shellrepeater is required within the
earthly-buildkit container.

This additionally removes the limitation of a single interactive
debugger per host, and allows attaching to remove buildkit instances.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>