Brew install should not fail if buildkit startup fails (#2172)
There are valid use cases for installing Earthly without a functional
Buildkit. For example, using it via podman (needs to be configured), or
using it only for org/account/secret management.

Before this PR, the brew installation would fail on a Mac if Docker is
not available. One case is that Docker simply is not installed. But
another (very common) case is that the Docker App is simply not started.
We don't want to prevent the user from installing Earthly in these
cases. We can always show them a more meaningful error on their first
build.

In addition, this PR improves the error shown when attempting to perform
a local build and Docker is not running.

Co-authored-by: Vlad A. Ionescu <vladaionescu@users.noreply.github.com>