Allow Feature Flags when Launching Satellites (#2165)
To enable a satellite with sleep/wake functionality:
```
earthly satellite launch --feature-flags satellite-sleep my-satellite
```

This can be re-used to enable feature flags we add for satellites in the
future as well. This is specific, however, to features that affect how
the satellite is created (vs other features which could take effect at
run-time during a build).