Support optional +secrets/ prefix (#2181)
In a previous change, we added a deprecation warning for the `+secrets/`
prefix but didn't actually make a code change to support its
optionality. This PR adds support for `NAME=SECRET` syntax without the
prefix. This should likely be allowed for all cases as it is not
particularly tied to cases that use the `PROJECT` feature.