Use gRPC Client Instead of JSON (#2187)
I was running into some issues with the JSON-to-protobuf unmarshaller
we've been using. In particular with enums, where existing enums can not
be added to on the server without breaking old clients versions - i.e.
the JSON unmarshaller throws an error if it encounters an enum value it
doesn't recognize. Traditional gRPC does not have this limitation.

Unsurprisingly, the code is also much simpler now in gRPC than it was
before using REST calls.

Note: this switches the Pipelines endpoints only, as other endpoints
don't support gRPC yet. In due time, we can convert the other endpoints
too (after we've converted them on the server first).