Make ./ast a submodule (#2190)
This PR moves `./ast` to a separate Go package. This makes it easier for
other libraries to import only the `ast` package.

I recall having used this approach in the past. The trade-off with
submodules like this is mostly around versioning. For example, it can
become a hassle to version bump the `ast` package as we develop the core
codebase. Not versioning the submodule independently is an option but
that can lead to surprises in dependent code. It's worth some
discussion.