Update crypto to address "parse error in message type 27" errors (#2280)
This (re)-fixes #2151 

After upgrading to 0.6.26 I started getting those errors again. It
appears that
https://github.com/earthly/earthly/commit/038239929f24ebdadcefc6b11c689a620b8dcda1
is the cause. The commit message/content doesn't indicate it was
intentional. The downgrade happens to be the exact version that I
upgraded from, which makes me think maybe it was a result of something
like a weird merge conflict resolution, external automation, etc.

Under the assumption it was not intentional, this is just the exact same
fix from #2152.