Satellite Wake-up Command (#2310)
The satellite wake command will start a satellite that is asleep. This
can be useful for using the `inspect` command, since it will not show
much info when the satellite is asleep.

The wake command will also use more "force" in trying to start a
satellite than a normal Reserve call will do, so it can be usef to debug
or attempt to recover a satellite that has failed or ended up in a bad
state for some reason.