Build ARM and x64 AMIs, move to Packer (#2335)
This makes it so we build an ARM and x64 Earthly AMI on release. I moved
from AWS ImageBuilder to this, because the build matrix was a little
weird keeping it in Terraform the way it was.

I tested the target called by `release.sh`, but not the full release
process.