Satellite sleep command (#2342)
Added a new command that manually puts a satellite to sleep. When the
satellite is already asleep, this is a no-op.

This command is very useful for us internally for testing, but also
could be of use to users who want more fine-grain control over their
satellite's up-time.

Note this will put a satellite to sleep, even if the user has manually
disabled the auto-sleep feature. Hence a user could opt out of the
auto-sleep feature and completely manage the satellite's state
themselves, if they are so inclined.