Improvements related to logstream integration (#2395)
Changes:

* Biggest change: Allow Earthly to run as a different
container+volume+.earthly dir name in development. This allows the
released Earthly to be logged in as one user and the dev Earthly to be
logged in as a different user. This works by changing the name `earthly`
in all these places to `earthly-dev`, when not building an official
release. So the container `earthly-buildkitd` becomes
`earthly-dev-buildkitd`, the cache volume `earthly-cache` becomes
`earthly-dev-cache`, and the dir `~/.earthly` becomes `~/.earthly-dev`.
The ports also shift. The release ports are the same as before, whereas
the ports used in dev are based on official ports + an offset calculated
based on a hash of the installation name.
* Removed debugger port as it was no longer used
* Add a flag for enabling uploading logstream: `--logstream-upload`.
When set, this replaces the old log uploading feature.
* Allow grpc insecure
* Remove unnecessary logstream batching

Co-authored-by: Vlad A. Ionescu <vladaionescu@users.noreply.github.com>