fix incorrect session ID passed to llb.SessionID
This fixes a bug where the incorrect session ID was passed to the copy
file context, this results in buildkit attempting to lock the main
session rather than the session associated with the build function. This
bug manifests itself as a time delay whenever copying a file. The time
delay occurs in buildkit due to a context timing out, which then falls back
to a different (available) session.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>