More fields on satellite ls and inspect (#2412)
Added a couple of fields to  `ls` and `inspect`. 

Buidlkit version and state are now shown in `ls`. Buildkit version only
works on new satellites, as it requires a new server change to be
applied as they are launched. Notice the output is also sorted now by
server too.

Feature flags are shown on `inspect`, along with the new version field.

Co-authored-by: Alex Couture-Beil <alex@earthly.dev>
Co-authored-by: Vlad A. Ionescu <446771+vladaionescu@users.noreply.github.com>