introduce a new flag for plain HTTP docker mirrors
434b98a8ad42d85883a6644bd22611678eadb101 removed the ability for a user
to run tests against a plain http-based mirror. This introduces a new
flag to allow configuring buildkit to use plain http.

Simply adding back the http option (and including the insecure option),
would result in the mirror being hit on both port 80 and por 443 (or if
a non-standard port was used, that port would be hit with both http and
https requests, which can lead to lots of noisy error logs).

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>