Allow registry login with cloud-stored secrets (#2597)
Allows a user to set registry credentials in the earthly-cloud secrets
system rather than `~/.docker/config.json`.

If a user is currently logged in to earthly-cloud, then earthly will use
credentials for the registry by searching:

1) /user/std/registry/<host>/username and
   /user/std/registry/<host>/password,

2) /org/project/std/registry/<host>/username and
   /org/project/std/registry/<host>/password,

and finally 3) the `~/.docker/config.json` file.

Note that all projects from both the directly and indirectly referenced
Earthfiles will be searched for authentication credentials. The
directly-referenced Earthfile will take prescedence, however there is no
guarentee the indirectly referenced Earthfiles will be consistently
ordered. This limitation means that it is not possible to associate a
particular `SAVE IMAGE --push ...` with a paticular org/project
credential storage location.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>