fix requirement of --org and --project flags
secrets v2 introduced the concept of a `--org` and `--project` flags. This
change deprecated the concept of global secrets at the org level; however,
this deprecation did not affect the /user/... secret namespace.

The /user namespace can still be accessed via `earthly secrets ls
/user/...`; however, the `--org` and `--project` flags were incorrectly
required all the time.

This commit changes the `--org` and `--project` flags to only be required
when the path is not prefixed by `/user/`.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>