fix: raise error when setting feature flag values
Boolean overrides for feature flags should not be possible; this change
will now raise an error instead of silently ignoring them.

While it is technically possible to implement disabling feature flags,
we chose to raise an error instead to encourage users to fully adopt
a new VERSION.

This is because feature flags are for experimental features, once they
are promoted they become part of the forward-direction of earthly;
allowing users to adopt the latest version while individually flipping a
feature off may lead to buggy experiences.

If a user really wants to disable a feature, then they will have to
remain at a previous VERSION, and instead select the (experimental) features
they wish to enable (with caution).

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>