fix release.sh not passing prerelease arg
The prerelease value was not correctly passed to the +release-github
target, which caused the RC to show up as latest.

This additionally re-orders the earthly version detection logic to
handle the case where the earthly command override is missing.

And checks that a tag containing a release candidate (rc) must be a
pre-release.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>