always apply SAVE ARTIFACT --keep-own flag
In order to copy a file from one target to another with the same
ownership information requires the `--keep-own` flag be applied to both
`SAVE ARTIFACT` and `COPY`.

The final `COPY`'s `--keep-own` flag should be sufficient for
determining if the original file's permissions are to be copied over or
not.

This fixes #2640

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>