logbus/logstreamer: use an iterator instead of channel
In order for us to solve the deadlock in the logstreamer, we
kind of need to use a different data structure than a channel.
As a first pass, we use an interface type as the contract
between the logstreamer and client packages, then implement
the interface using channels.

While working through this implementation, we also found a
data race and added a test to prove it. Similar to the
deadlock, it requires scheduler luck.