cloud.Client should just be a struct
As per https://github.com/golang/go/wiki/CodeReviewComments#interfaces
an interface should be defined wherever it is used, and not in the
cloud package.

If an interface is needed by a specific caller of the client, then it's
best to create a much smaller interface tied to the use of the client,
which will make it more clear which methods must bet implemented for
that specific portion of the code.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>