logbus/logstreamer: fix drain blocking on read from nil channel (#2686)
I also added a test for `Close()` draining when it's being flooded with
messages, because I think it would have issues with that without the
mutex lock (and since it's a channel, I wouldn't be surprised if the
mutex lock is removed some day). However, due to the mutex lock, it's
actually not a current problem. But having a test there for if/when the
mutex lock is removed is nice.