ast: handle unindented comments in recipe bodies
This solves an issue where unindented comments in recipe bodies would
trigger a DEDENT token, causing parser errors when the comments really
should just be ignored.

This also incidentally allows documentation to be on a line immediately
following a comment if that comment is in an indented block (e.g. if
the previous target ended with a comment).