Update base image for scala example (#2733)
The scala example did not work on my MacBook Pro M2 Max running colima.
The process was stalled indefinitely (or at least over 1 hour). Memory
was not the issue, but the base image.
Just running a compilation with `docker run -it --rm -v $PWD:/work -w
/work hseeberger/scala-sbt:11.0.7_1.3.13_2.11.12 sbt package` also did
not complete.
Using the `scala-sbt` as base image did work.