variables: rewrite Scope with functional args
Instead of using 'scope.GetActive()' and 'scope.GetInactive()' (and all of the other variations),
use 'scope.Get(variables.WithActive())' and the like. This gives us more flexibility for setting
explicit environment variables and such in the future.

This also gives us an option for rejecting overrides, which is useful for ARG scope and SET.