Use source filename as destination if unspecified
Makes `SAVE ARTIFACT foo AS LOCAL dir/` export the result to `dir/foo`.  The file receiver does not have the source information to do it itself.

This change requires an explicit whitelisting.  Given a working directory of `/tmp`, a `SAVE ARTIFACT foo AS LOCAL docs/` would previously result in the converter adding `/tmp/foo` to the whitelist, and waitBlock adding `docs/`.  The file receiver (now) gets `docs/foo`, which wasn't whitelisted.  In a scenario where the user gave a destination of `foo`, the entries would be `/tmp/foo` (from converter) and `foo` (from waitBlock), with the receiver getting `foo`.

In other words, the `waitBlock` was the only thing providing legitimate whitelist entries for this "save on failed `TRY`" flow, which was probably not the intent.  The current converter code always whitelists an absolute path, but the file receiver is always given a relative path.