Add command line options for time/size based partial cache pruning (#2822)
This adds two command line options for pruning the cache.

* `--keep-duration`, which will drop all cache older than this duration.
It expects a string parseable by `time.ParseDuration`.
* `--target-size`, which will drop the oldest items from the cache until
the size is reached. It expects a string parseable by
`humanize.ParseBytes` (which we already have in our `go.mod`).

I presume docs need to be added, I'll append those to this PR if these
options look ok.