Buildkit Entrypoint - PIDs excludable from OOM reaper (#2870)
Added an extension to the Buildkit entrypoint which allows the caller to
specify one or more PIDs that cannot be OOM-killed.

The OOM-reaper did not normally allow extra PIDs to exist in the
container, because it assumed they were "ghost" processes left lingering
after something was OOM-killed. This made it difficult to incorporate
our new "runner" process which runs now on Satellites and CI runners.