loosen the tests required before doing a pre-release
Due to flaky tests, our pre-release workflow is not running; this means
we are no longer dogfooding our code in main via the ./earthly script.

If the code is merged into main; then we need to be dogfooding it --
even if it causes tests to fail -- otherwise we won't ever taste bad
dogfood.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>