Initialize log streaming upfront if org/project envs set (#2892)
This PR will help ensure that logs are (almost) always sent to the log
streaming service when streaming is enabled. It does so by looking for
`EARTHLY_ORG` and `EARTHLY_PROJECT` and initializing the stream using
those values when present. The remote end will verify that the currently
logged-in user has permission to write logs for the given org/project
combination.