git-lfs support for remote targets
Experimental support for git lfs when referencing a remote target.

Usage:

  earthly --git-lfs-pull-include=<path> github.com/user/project+target

This will perform a `git lfs pull --include=<path>` in the root of the
repo, prior to running the +target.

This is a (hopefully) temporary fix for https://github.com/earthly/earthly/pull/2922

Note that this does not work with the `GIT CLONE` command.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>