.github: add a workflow to sync issue state from the board
We already have (built-in) workflows in our project that will update issues' place
on the board when their state changes. Closed issues will move to Done, reopened
issues will move to Todo.

However, the opposite is not true. When we move an issue to Done, it will not be
automatically closed.

This workflow solves that by polling the board (once per hour) and updating any
tickets that have been moved so that their state matches their place on the board.