prevent auto-ssh login after explicit logout
Earthly, by default, will attempt to auto-login whenever
`SSH_AUTH_AGENT` is running and has keys -- this is confusion
for users who do an explicit `earthly account logout`.

Now when an explicit logout is performed, the file
`~/.earthly/do-not-login-automatically` will be created, which will
prevent auto logins in the future.

This file is deleted when a user performs an explicit `earthly account
login ...` .

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>