Remove default size in satellite launch (#3057)
This will now be determined by the backend when size is empty. The
defaults are small for free tier, medium for paid tiers.

It is especially important to remove this for free tier users, as it
will cause errors with it set to medium by default since their plan
doesn't allow that size.