.github: spread out multiple test targets across multiple workflows (#3024)
This splits up some of our longest-running (and most likely to flake)
GHA jobs into smaller chunks. This helps specifically in the case where
a test flakes, requiring us to re-run the job - taking around 10 minutes
each retry instead of 45.

Hopefully, this will also give us a clearer picture of which tests are
flaking most often, so we know which ones need to be fixed.