authprovider: invalidate cache when AddProject is called
Part of the improvements in PR #3026 was a form of caching authproviders that don't have
any credentials configured for a given host. This did not account for the project being
set *after* the first attempt at loading credentials.

This commit updates the aggregated authprovider to invalidate its cache of per-host
child auth providers when AddProject is called.