use acbgrep to make test failures more clear
`acbgrep` from https://github.com/alexcb/acbtest; is a wrapper around
grep which will display a failure message if grep fails (rather than
silently exiting with an error code).

When data is pipped in, it will also display the contents of the data
that is being grepped -- this makes it easier to see the full output of
earthly if grep fails to find an expected string.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>