print summary of interactive debugger tests
Display a summary of the total number of tests that passed, and display
which ones failed at the end of the test runner.

Without this, a user might have a previous test fail, but the final test
passes and displays:

    /test # cat /data.txt | base64 -d
    e88cc2b8-c179-4ed7-8eae-207a0ef7546c/test # [6n
    test-docker-compose passed
    Error: Process completed with exit code 1.

which is confusing, as they might not know to scroll up looking for a
previous test failure.

Now we will display:

    == Summary ==
    2/3 tests passed
    test-interactive-run failed with exit code=1

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>