Selected org in config works with secrets (#3112)
The `org select` command now propagates to other earthly commands for
`secrets`, `registry`, and `project`.

The secrets needed some additional tweaking to support the selected org.
That's because we always assumed that user-space secrets would be
addressed and displayed with a `/user` prefix. Now that we can select a
personal org, user-space secrets can be addressed and displayed without
the `/user` prefix. The prefix is still supported though, and works even
when a different org is selected.