prompt for secret value when missing
When running `earthly secret set <path>`, rather than exiting with an error,
we will prompt the user for a single-line secret value.

This allows a user to type in a single line followed by hitting the
<enter> key.

Additionally, if `-stdin` is used from an interactive tty, earthly will
now display a message saying it is waiting for an EOF (ctrl-d).

This additionally fixes a bug with isTTY, which was only checking stdout
but could not detect if a command was being piped to stdin or not.

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>