Fix account create-token help string (#3250)
The help string says the tokens will expire in one year by default,
however, the earthly code (and corresponding cloud code) ssem to be
deliberately using no expiry as the default.

FWIW I think no-expiry as a default makes sense, as these tokens will
often be used in CI for satellites. It may result in an unpleasant
surprise a year later if the token expires and the user's CI fails.