Buildkit entrypoint kills externally provided pids (#3328)
The current `OOM_EXCLUDED_PIDS` variable allows one or more process IDs
to be passed-in externally to buildkit, so that its oom adj script will
not attempt to reap them. This is currently used by the runner service
(i.e. satellites) so that the runner process can run alongside buildkit
in a container.

This PR allows buildkit to also gracefully kill those pids as part of
its own shutdown process, which in turn allows our runner service to
gracefully shutdown as well.

---------

Co-authored-by: Alex Couture-Beil <alex@earthly.dev>