Registry proxy: Use lower-level TCP streaming (#3317)
Removes the HTTP-level code in favor of a lower-level TCP proxy
approach.

The tricky part here was that Docker (and most HTTP clients) leaves the
TCP connection open. Hence, there's no way for the code to detect when
to stop reading the initial request aside from a timeout. Apparently,
the recommended approach is to set
[`SetReadDeadline`](https://cs.opensource.google/go/go/+/refs/tags/go1.21.1:src/net/net.go;l=156)
ahead of each read and appropriately handle any timeout errors.

Overall, I think this is preferable to the original code as the HTTP
code & header parsing is a bit brittle.

Works with: https://github.com/earthly/buildkit/pull/26