Clarify errors when attempting to pass secrets to a build (#3500)
This PR will clarify error messages when there is a problem setting
secrets in a build command.
It includes a refactor the the hintError so that hints will be printed
more consistently when handling errors at the final of the binary
execution.

I didn't want to change how all hints are printed as part of this PR
(will take too long atm to confirm how these messages will be displayed)
so I created deprecated methods to create hint errors so that other
hints will still be displayed in the same way they have so far.